% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapas.R
\name{kml_cortar}
\alias{kml_cortar}
\title{Reduce un KML a otra geografia}
\usage{
kml_cortar(obj, condicion)
}
\arguments{
\item{obj}{Objeto SpatialPolygonsDataFrame}

\item{condicion}{condicion bajo cual filtrar}
}
\description{
Hace un subset de una geografia, para revisar una zona particular de un KML.
}

