% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapas.R
\name{t_expkml}
\alias{t_expkml}
\title{Exportar KML con rangos de colores}
\usage{
t_expkml(obj, archivo, col_nombre, col_desc, col_color,
  mapa_colores = "mapa_colores_topaz", mapa_nombre = "Mapa_TSE",
  mapa_desc = "Mapa Creado por PE de TSE")
}
\arguments{
\item{obj}{objeto SpatialPolygonsDataFrame, pasado como string.}

\item{archivo}{nombre de archivo al que se va exportar.}

\item{col_nombre}{columna en data.frame con nombre de cada objeto.}

\item{col_desc}{columna con descripcion en data.frame para cada objeto.}

\item{col_color}{columna con id's de color en mapa de colores.}

\item{mapa_colores}{data.frame con columna de id y color para mapear los id's a un color #RRGGBBAA form (AA=alpha 00=transparent FF=solid). Defaults a: "mapa_colores_topaz".}

\item{mapa_nombre}{nombre de mapa (capa) que se visualiza en Google Earth. Defaults a: "Mapa_TSE".}

\item{mapa_desc}{descripcion del mapa (capa) que se visualiza en Google Earth. Defaults a: "Mapa creado por PE de TSE".}
}
\description{
Exporta un archivo KML con atributos de colores, para usar en Google Maps. Usa la funcion de Tom Philippi
(disponible en: https://dl.dropboxusercontent.com/content_link/DcDBdbghu5D7eM6tBQ5cVGMNGYXaSzaj3CZfcYSc9vjUDW3yd3KrTynMotNZoi4n/file)
Exporta en loop el objeto SpatialPolygonsDataFrame a un KML. Todos los parametros se pasan como strings.
}

