% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapas.R
\name{t_impkml}
\alias{t_impkml}
\title{Importar KML con Codigos o AGEBS}
\usage{
t_impkml(archivo)
}
\arguments{
\item{archivo}{nombre de archivo}
}
\description{
Importa en R un KML exportado por entidad oficial, para usarse en crear mapas.
}

