% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema.R
\name{tse_colores}
\alias{tse_colores}
\title{Exporta paletas de colores}
\usage{
tse_colores(paleta)
}
\arguments{
\item{paleta}{paleta de colores, en numero (1).}
}
\description{
Escoges una paleta y exporta colores correspondientes.
}
\details{
Paleta 1: creada por KFG (25-01-2017)
}

