% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema.R
\name{tse_theme}
\alias{tse_theme}
\title{Funcion generica para temas}
\usage{
tse_theme(tipo = "hc")
}
\arguments{
\item{tipo}{tipo de grafica (hc, ggplot, o plotly)}

\item{colores}{esquema de colores, generalmente el default (tse1)}
}
\description{
Temas para graficas (ggplot, ploty o highcharter)
}

